Redirect Data Processing workshop to web version
Updates to this workshop have been done outside of this repo and the
content within this repo is badly out of date. I'd like to nuke the
content here for the time being. If there's interest in maintaining
this repository, we can backport the contents from the web version in
a separate work effort.

Closes #86, closes #91, closes #125